Moves parent category validator
Moves the parent category validator service to a dedicated 'validators' directory.

This change improves the project's structure and enhances code organization by grouping validators in a single directory.
