+++ src/modules/categories/categories.module.ts
@@ -6,7 +6,7 @@ import { CategoriesService } from './categories.service';
 import { CategoriesController } from './categories.controller';
 import { Category } from './entities/category.entity';
 import { ParentCategoryFinder } from './finders/parent-category.finder';
-import { ParentCategoryValidatorService } from './parent-category.validator';
+import { ParentCategoryValidatorService } from './validators/parent-category.validator';
 import { CategoryValidatorService } from './category.validator.service';
 
 @Module({
