+++ src/modules/categories/dto/update-category.dto.ts
@@ -1,7 +1,7 @@
 import { PartialType } from '@nestjs/swagger';
 import { CreateCategoryDto } from './create-category.dto';
 import { IsNotEmpty, IsOptional, IsString, IsBoolean, IsUUID, Validate } from 'class-validator';
-import { ParentCategoryValidatorService } from '../parent-category.validator';
+import { ParentCategoryValidatorService } from '../validators/parent-category.validator';
 // export class UpdateCategoryDto extends PartialType(CreateCategoryDto) {}
 
 export class UpdateCategoryDto {
