+++ src/modules/categories/validators/parent-category.validator.ts
@@ -1,8 +1,8 @@
 import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import { Category } from "./entities/category.entity";
+import { Category } from "../entities/category.entity";
 import { Repository } from "typeorm";
-
+Category
 @Injectable()
 export class
     ParentCategoryValidatorService {
